% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLRinfo.R
\name{filterLRinfo}
\alias{filterLRinfo}
\title{Filter LR info (expression, interaction score, and ligand score)}
\usage{
filterLRinfo(LRinfolist, cut.bounds)
}
\arguments{
\item{LRinfolist}{output of "getLRinfosummary".}

\item{cut.bounds}{a list of lower- and upper-bounds of selected columns of data.frames in LRinfolist.}
}
\value{
Filtered LRinfolist
}
\description{
Filter LR info (expression, interaction score, and ligand score)
}
