% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircosPlot.R
\name{getCircosPlot}
\alias{getCircosPlot}
\title{Create circos plot of identified LR interactions}
\usage{
getCircosPlot(
  plotLRinfolist,
  plotcond,
  cls_from,
  cl_to,
  mycols,
  width_same_ligand_cluster,
  width_different_ligand_cluster,
  width_ligand_receptor,
  width_same_receptor_cluster,
  width_different_receptor_cluster,
  cplotthresh,
  cex,
  transupperbase,
  transidx
)
}
\arguments{
\item{plotLRinfolist}{the "LRinfolist_fil" entry of the output from the function "filterLRinfo".}

\item{plotcond}{specifies from which condition the interaction scores will be plot,"condmax" (the maximum interaction score across conditions) or one of the conditions under comparison.}

\item{cls_from}{sender cell-types/clusters.}

\item{cl_to}{specifies receiver cell-type/cluster.}

\item{mycols}{a vector of colors for related cell-types/clusters.}

\item{width_same_ligand_cluster}{gap width between ligand nodes in the same cluster.}

\item{width_different_ligand_cluster}{gap width between different ligand clusters.}

\item{width_ligand_receptor}{gap width between the ligand nodes and the receptor nodes.}

\item{width_same_receptor_cluster}{gap width between receptor nodes in the same cluster.}

\item{width_different_receptor_cluster}{gap width between different receptor clusters.}

\item{cplotthresh}{specifies threshold of the SLRI values in the plot. Only plot LR interactions with SLRI > cplotthresh.}

\item{cex}{font size.}

\item{transupperbase}{parameter of edge transparency adjustment.}

\item{transidx}{parameter of edge transparency adjustment.}
}
\value{
Circos plot of identified LR interactions
}
\description{
Create circos plot of identified LR interactions
}
