% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExprInfo.R
\name{getExprInfo}
\alias{getExprInfo}
\title{Get gene average expression, detection rate, and differential expression info}
\usage{
getExprInfo(seuratobj, cls, conds, condpairs, testmethod = "wilcox")
}
\arguments{
\item{seuratobj}{Seurat object of the input scRNA-seq data, with sample conditions and cell-types/clusters in the columns "cond" and "cl" of the meta.data.}

\item{cls}{cell-types/clusters.}

\item{conds}{sample conditions.}

\item{condpairs}{a matrix with two columns "cond1" and "cond2", each row is ordered, represents a comparison between two sample conditions of interest, cond1 vs cond2 in differential expression analysis.}

\item{testmethod}{the test.use option for differential expression analysis in the function "FindMarkers".}
}
\value{
A list exprinfo:
exprinfo$avgexprall: the average expression of all genes in all cells
exprinfo$avgexprlist: list of average expression matrix (genes in rows, sample conditions in columns) of each cell-type/cluser in cls
exprinfo$pctexprlist: list of gene detection rate matrix (genes in rows, sample conditions in columns) of each cell-type/cluser in cls
exprinfo$DElist: list of differential expression matrix (genes in rows) of each cell-type/cluser in cls, each sample condition pair in condpairs
}
\description{
Get gene average expression, detection rate, and differential expression info
}
