% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKEGGNESbarplot.R
\name{getKEGGNESbarplot}
\alias{getKEGGNESbarplot}
\title{get KEGG pathway NES barplot}
\usage{
getKEGGNESbarplot(
  mysig,
  padj.cut = 0.25,
  num.decimal = 2,
  break.bound = 2,
  break.step = 1
)
}
\arguments{
\item{mysig}{signature list of input data.}

\item{padj.cut}{specifies cutoff value on the adjusted p.values of the pathways from GSEA.}

\item{num.decimal}{number of decimal places to keep for the NES values in the plot.}

\item{break.bound}{breaks on the x-axis is set to seq(break.bound*(-1), break.bound, by = break.step).}

\item{break.step}{breaks on the x-axis is set to seq(break.bound*(-1), break.bound, by = break.step).}
}
\value{
barplot of the NES values of enriched KEGG pathways (with adjusted p.value < padj.cut) resulted from GSEA
}
\description{
get KEGG pathway NES barplot
}
