% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLRIScore.R
\name{getLRIScore}
\alias{getLRIScore}
\title{Score LR interactions in specified lr_network in each sample condition}
\usage{
getLRIScore(lr_network, exprinfo, LRI.method = "scsigr")
}
\arguments{
\item{lr_network}{a ligand_receptor network matrix with columns "L", "Lgene", "R", and "Rgene".}

\item{exprinfo}{output of function "getExpInfo".}

\item{LRI.method}{"cpdb", "natmi", or "scsigr".}
}
\value{
List of LR interaction score and score-specificity data.frames for each cell type/cluster-pair in each sample condition
}
\description{
Score LR interactions in specified lr_network in each sample condition
}
