% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLScore.R
\name{getLScore}
\alias{getLScore}
\title{Score the ligands in Lignature}
\usage{
getLScore(siggroups, lr_network, sigscorelist, pvalcut = 1)
}
\arguments{
\item{siggroups}{list of signatureid-vectors for each ligand in Lignature.}

\item{lr_network}{a ligand_receptor network matrix with ligand names in the column "L".}

\item{sigscorelist}{output from the function "getSigScore".}

\item{pvalcut}{p-value cutoff, only signatures with score-pvalues below this cutoff are used for scoring the ligands in Lignature.}
}
\value{
a data.frame of ligand scores and the signature sources
}
\description{
Score the ligands in Lignature
}
