% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMySigs.R
\name{getMySigs}
\alias{getMySigs}
\title{Get signatures of input data}
\usage{
getMySigs(
  DElist,
  cls,
  condpairs,
  keggnodes,
  pctcut = 0.1,
  belowpctcut = "as.is",
  testdegpadjcut = 1
)
}
\arguments{
\item{DElist}{output of function "getExpInfo", exprinfo$DElist, a list of differential expression matrix (genes in rows) of each cell-type/cluser in cls, each sample condition pair.}

\item{cls}{cell-types/clusters.}

\item{condpairs}{a matrix with two columns "cond1" and "cond2", each row is ordered, represents a comparison between two sample conditions of interest, cond1 vs cond2 in differential expression analysis.}

\item{keggnodes}{list of the gene-nodes (represented by gene symbols) of KEGG pathways.}

\item{pctcut}{gene detection rate cutoff-value.}

\item{belowpctcut}{specify how to deal with avg_log2FC values of genes with detection rate below pctcut, one of "as.is", "lowpct.as.zero", or "lowpct.remove".}

\item{testdegpadjcut}{gene p_val_adj (from differential expression analysis) cutoff-value, avg_log2FC of genes with p_val_adj above this value is zeroed out.}
}
\value{
a list of signatures for each cell-type/cluser in cls, each sample condition pair in condpairs
}
\description{
Get signatures of input data
}
