% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathNodes.R
\name{getPathNodes}
\alias{getPathNodes}
\title{Get gene-nodes in KEGG pathways}
\usage{
getPathNodes(
  species = "hsapiens",
  database = "kegg",
  numnode_min = 10,
  numnode_max = 500
)
}
\arguments{
\item{species}{one of the supported species.}

\item{database}{the name of the pathway database.}

\item{numnode_min}{minimum number of gene-nodes in the KEGG pathways to be collected (KEGG pathways with fewer number of gene nodes are removed).}

\item{numnode_max}{maximum number of gene-nodes in the KEGG pathways to be collected (KEGG pathways with larger number of gene nodes are removed).}
}
\value{
a list of the gene-nodes (represented by gene symbols) of KEGG pathways with the number of included gene-nodes bounded by numnode_min and numnode_max
}
\description{
Get gene-nodes in KEGG pathways
}
