% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScatter.R
\name{getScatter}
\alias{getScatter}
\title{For ligands in identified LR interactions, create scatterplot of the ligand scores and the maximum avgexpr levels acorss all sender cells and conditions.}
\usage{
getScatter(
  LRinfolist,
  LRs,
  cls_from,
  cl_to,
  condpair,
  Lscore.by = "Lscore_max",
  point.size = 3
)
}
\arguments{
\item{LRinfolist}{output of "getLRinfosummary".}

\item{LRs}{vector of identified LR interactions "L_Lgene_R_Rgene".}

\item{cls_from}{sender cell-types/clusters.}

\item{cl_to}{specifies receiver cell-type/cluster.}

\item{condpair}{sample condition pair for the plot, in the form of "cond1_vs_cond2".}

\item{Lscore.by}{how to score each ligand with respect to the scores of the set of signatures corresponding to it, one of "Lscore_mean", "Lscore_median", "Lscore_max", "Lscore_min", "Lscore_max_abs", and "Lscore_signed_maxabs".}

\item{point.size}{point size.}
}
\value{
Scatterplot of the ligand scores and the maximum avgexpr levels acorss all sender cells and conditions for ligands in identified LR interactions.
}
\description{
For ligands in identified LR interactions, create scatterplot of the ligand scores and the maximum avgexpr levels acorss all sender cells and conditions.
}
