% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScatterLRpairs.R
\name{getScatterLRpairs}
\alias{getScatterLRpairs}
\title{Get scatter plot of specified ligand scores and LR-interactoin scores (maximum from all sender cells) for LR pairs with these scores available.}
\usage{
getScatterLRpairs(
  LRinfolist,
  cls_from,
  cl_to,
  Lscore.by = "Lscore_max",
  SLRI.by = "SLRI_condmax",
  point.size = 3,
  Lscore.cut = 0.1,
  SLRI.cut = 0.3
)
}
\arguments{
\item{LRinfolist}{output of "getLRinfosummary".}

\item{cls_from}{sender cell-types/clusters.}

\item{cl_to}{specifies receiver cell-type/cluster.}

\item{Lscore.by}{how to score each ligand with respect to the scores of the set of signatures corresponding to it, one of "Lscore_mean", "Lscore_median", "Lscore_max", "Lscore_min", "Lscore_max_abs", and "Lscore_signed_maxabs".}

\item{SLRI.by}{specifies which LR-interaction score to plot, e.g. "SLRI_condmax" and "SLRIspec_condmax".}

\item{point.size}{point size.}

\item{Lscore.cut}{ligand score threshold.}

\item{SLRI.cut}{LR-interaction score threshold.}
}
\value{
Scatterplot of specified ligand scores and LR-interactoin scores (maximum from all sender cells) for LR pairs with these scores available.
}
\description{
Get scatter plot of specified ligand scores and LR-interactoin scores (maximum from all sender cells) for LR pairs with these scores available.
}
