% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSigScore.R
\name{getSigScore}
\alias{getSigScore}
\title{Score each signature in Lignature given signatures of input data}
\usage{
getSigScore(
  siglist,
  testsig,
  whichsig = "lfc",
  sigcut = 0.1,
  belowsigcut = "as.is",
  padjcut = 1,
  minsiglength = 10,
  method = "cor.pearson",
  nump = 1000,
  corpby.permu = FALSE
)
}
\arguments{
\item{siglist}{siglist of Lignature.}

\item{testsig}{input signatures.}

\item{whichsig}{"lfc" or "NES".}

\item{sigcut}{cutoff value applied to the signatures.}

\item{belowsigcut}{"as.is", "remove", or "as.zero", specify how to deal with signature values below sigcut.}

\item{padjcut}{cutoff values applied to the adjusted p.values of each entry of the signatures.}

\item{minsiglength}{minimum length of signatures to compare.}

\item{method}{similarity measurement metric, one of "cor.pearson", "cor.spearman", "cosine.similarity", "euclidean", "manhattan", "euclidean_mean", "manhattan_mean".}

\item{nump}{number of permutations of the input test signature in the calcuation of score-pvalues.}

\item{corpby.permu}{if TRUE, score-pvalues represent results from permutations when method is set to "cor.pearson" or "cor.spearman".}
}
\value{
a list of two vectors: a vector of signature scores and a vector of score-pvalues
}
\description{
Score each signature in Lignature given signatures of input data
}
