% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSigScore.R
\name{mymetric}
\alias{mymetric}
\title{Define similarity metrics}
\usage{
mymetric(a, b, mymethod)
}
\arguments{
\item{a}{vector a}

\item{b}{vector b}

\item{mymethod}{similarity metric}
}
\value{
similarity score
}
\description{
Define similarity metrics
}
