% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDELRheatmap.R
\name{plotDELRheatmap}
\alias{plotDELRheatmap}
\title{Plot expression of differentially expressed ligand or receptor genes in each cell-type/cluster, each condition, in heatmap}
\usage{
plotDELRheatmap(
  exprinfo,
  siggroups,
  lr_network,
  whichLs = "in.Lignature.only",
  cls,
  condpair,
  LorR = "L",
  pct.cut = 0.1,
  padj.cut = 0.05,
  lfc.cut = 0.1,
  col,
  scale = "row",
  cluster_cols = F,
  cluster_rows = T,
  show_rownames = T,
  treeheight_row = 0,
  treeheight_col = 0,
  annotation_legend = TRUE,
  border_color = FALSE,
  fontsize_row = 13,
  fontsize_col = 13,
  fontsize = 13
)
}
\arguments{
\item{exprinfo}{output of function "getExpInfo".}

\item{siggroups}{list of signatureid-vectors for each ligand in Lignature.}

\item{lr_network}{ligand_receptor network matrix with columns "L", "Lgene", "R", and "Rgene".}

\item{whichLs}{specifies the set of ligand genes to examine. "in.lrnetwork" (all ligand genes in lr_network) or "in.Lignature.only" (ligand genes in lr_network that are included in Lignature).}

\item{cls}{cell-types/clusters to plot.}

\item{condpair}{sample condition pair to plot, in the form of "cond1_vs_cond2".}

\item{LorR}{plot expression of differentially expressed ligand (L) or receptor (R) genes.}

\item{pct.cut}{cutoff value on the maximum detection rate across the sample conditions under comparison for each gene, for the identification of differentially expressed genes.}

\item{padj.cut}{cutoff value on the adjusted p.values of genes in the differential expression analysis, for the identification of differentially expressed genes.}

\item{lfc.cut}{cutoff value on the |avg_log2FC| values of genes in the differential expression analysis, for the identification of differentially expressed genes.}

\item{col}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{scale}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{cluster_cols}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{cluster_rows}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{show_rownames}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{treeheight_row}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{treeheight_col}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{annotation_legend}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{border_color}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{fontsize_row}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{fontsize_col}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{fontsize}{same as in the function "pheatmap" in the "pheatmap" package.}
}
\value{
heatmap of the average expression values of differentially expressed ligand or receptor genes in each cell-type/cluster in each condition
}
\description{
Plot expression of differentially expressed ligand or receptor genes in each cell-type/cluster, each condition, in heatmap
}
