% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLRs.R
\name{plotLRs}
\alias{plotLRs}
\title{Plot the average expression of ligands in sender cells, receptors in specified receiver cells, or the ligand scores calculated by Lignature signatures for identified LR interactions}
\usage{
plotLRs(
  LRinfolist,
  LRs,
  cls_from,
  cl_to,
  condpair,
  Lscore.by = "Lscore_max",
  order.by,
  order.decreasing = T,
  plotwhich = "LScore",
  col,
  scale = "none",
  cluster_cols = F,
  cluster_rows = F,
  show_rownames = T,
  treeheight_row = 0,
  treeheight_col = 0,
  annotation_legend = TRUE,
  border_color = FALSE,
  fontsize_row = 13,
  fontsize_col = 13,
  fontsize = 13
)
}
\arguments{
\item{LRinfolist}{output of "getLRinfosummary".}

\item{LRs}{vector of LR interactions "L_Lgene_R_Rgene" to plot.}

\item{cls_from}{sender cell-types/clusters.}

\item{cl_to}{specifies receiver cell-type/cluster.}

\item{condpair}{sample condition pair to plot, in the form of "cond1_vs_cond2".}

\item{Lscore.by}{how to score each ligand with respect to the scores of the set of signatures corresponding to it, one of "Lscore_mean", "Lscore_median", "Lscore_max", "Lscore_min", "Lscore_max_abs", and "Lscore_signed_maxabs".}

\item{order.by}{order the LR interactions by which column of the matrices in LRinfolist.}

\item{order.decreasing}{TRUE or FALSE, whether sort the LR interactions in a descending order or not, according to the values in the column specified by "order.by".}

\item{plotwhich}{plot average expression of ligand ("Lavgexpr") or receptor ("Ravgexpr") genes, or the ligand scores specified by "Lscore.by".}

\item{col}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{scale}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{cluster_cols}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{cluster_rows}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{show_rownames}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{treeheight_row}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{treeheight_col}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{annotation_legend}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{border_color}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{fontsize_row}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{fontsize_col}{same as in the function "pheatmap" in the "pheatmap" package.}

\item{fontsize}{same as in the function "pheatmap" in the "pheatmap" package.}
}
\value{
Heatmap of the average expression of ligands in sender cells in each condition, average expression of receptors in specified receiver cells in each condition,
or the corresponding ligand scores calculated by Lignature signatures, for identified LR interactions
}
\description{
Plot the average expression of ligands in sender cells, receptors in specified receiver cells, or the ligand scores calculated by Lignature signatures for identified LR interactions
}
