% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLRinfosummary.R
\name{signedmaxabs}
\alias{signedmaxabs}
\title{get signed values}
\usage{
signedmaxabs(x, y)
}
\arguments{
\item{x}{absolute value vector.}

\item{y}{sign vector.}
}
\value{
signed values
}
\description{
get signed values
}
