% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePathSigs.R
\name{updatePathSigs}
\alias{updatePathSigs}
\title{Update KEGG pathway signatures from fgsea}
\usage{
updatePathSigs(siglist, keggnodes, progress_interval = 10)
}
\arguments{
\item{siglist}{the current siglist of Lignature.}

\item{keggnodes}{list of the gene-nodes (represented by gene symbols) of KEGG pathways.}

\item{progress_interval}{report progress very progress_interval runs in the for-loop.}
}
\value{
siglist with updated pathway signatures
}
\description{
Update KEGG pathway signatures from fgsea
}
