% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLRtable.R
\name{writeLRtable}
\alias{writeLRtable}
\title{Write a list of data.frames in .csv tables}
\usage{
writeLRtable(tablelist, mydir)
}
\arguments{
\item{tablelist}{a list of data.frames.}

\item{mydir}{the directory where the tables will be saved.}
}
\value{
.csv tables saved in mydir
}
\description{
Write a list of data.frames in .csv tables
}
